requirements.txt
# Byzantine-Robust Identity Framework - Dependencies

# Core ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Blockchain
web3>=6.0.0

# Reports
reportlab>=4.0.0

# Dashboard
streamlit>=1.28.0

# Config
pyyaml>=6.0

# Utilities
plotly>=5.0.0
matplotlib>=3.5.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
